pyproject.toml
[project]
name = "full-stack-automation-cyera"
version = "0.1.0"
description = "Pytest + Playwright automation framework for the Cyera full-stack home assignment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "allure-pytest>=2.13.5",
    "assertpy>=1.1",
    "playwright>=1.53.0",
    "pytest>=8.4.0",
    "pytest-playwright>=0.7.0",
    "requests>=2.32.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["api*", "config*", "core*", "ui*"]
